pyproject.toml
[build-system]
requires = [
    "flit_core >=3.2,<4",
    "python_version >= '3.7'"
]
build-backend = "flit_core.buildapi"

[project]
name = "align"
authors = [{name = "Giovanni Aviles", email = "giovanni.aviles@ucsf.edu"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "pytest",
    "numpy",
    "biopython"
]

[project.urls]
Home = "https://github.com/ggaviles/HW5-NW/tree/main/align"
